requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
httpx>=0.24.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
pandas>=2.1.1
pydantic>=2.2.1
google-genai>=0.1.1
python-dotenv>=1.0.0